refactor:扣积分重构

See merge request miotech-application/backend/mp2c-go!311

+++ internal/pkg/model/entity/point_transaction.go
@@ -155,7 +155,7 @@ type PointTransaction struct {
 	OpenId         string               `gorm:"column:openid" json:"openId"`
 	TransactionId  string               `json:"transactionId"`
 	Type           PointTransactionType `json:"type"`
-	Value          int                  `json:"value"`
+	Value          int64                `json:"value"`
 	CreateTime     model.Time           `json:"createTime"`
 	AdditionalInfo AdditionalInfo       `json:"additionalInfo"`
 	AdminId        int                  `json:"adminId"`
